+++ app/services/answer_factory.rb
@@ -1,0 +1,15 @@
+class AnswerFactory
+  attr_accessor :question
+
+  def initialize(question:)
+    self.question = question
+  end
+
+  def call
+    case question.contentful_type
+    when "radios" then RadioAnswer.new
+    when "short_text" then ShortTextAnswer.new
+    else Answer.new
+    end
+  end
+end