+++ spec/models/short_text_answer_spec.rb
@@ -1,0 +1,14 @@
+require "rails_helper"
+
+RSpec.describe ShortTextAnswer, type: :model do
+  it { should belong_to(:question) }
+
+  it "captures the users response as a string" do
+    answer = build(:short_text_answer, response: "Yellow")
+    expect(answer.response).to eql("Yellow")
+  end
+
+  describe "validations" do
+    it { is_expected.to validate_presence_of(:response) }
+  end
+end