+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog 1.0.0].
 
 ## [Unreleased]
 
+## [release-002] - 2020-11-16
+
+- migrate answer database table into Radio and ShortText
+
 ## [release-001] - 2020-11-12
 
 - address rails-template TODO
@@ -22,6 +26,7 @@ Contentful fixture
 - Contentful can redirect users to preview endpoints
 
 [unreleased]: https://github.com/DfE/DFE-Digital/buy-for-your-school/compare/release-001...HEAD
+[release-002]: https://github.com/DFE-Digital/buy-for-your-school/compare/release-001...release-002
 [release-001]: https://github.com/DFE-Digital/buy-for-your-school/compare/release-000...release-001
 
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/